+++ src/backend/lib/redis.js
@@ -1,8 +1,8 @@
-require('../config');
+require('./config');
 const Redis = require('ioredis');
 const MockRedis = require('ioredis-mock');
 
-const { logger } = require('./logger');
+const { logger } = require('../utils/logger');
 
 // If you need to set the Redis URL, do it in REDIS_URL
 const redisUrl = process.env.REDIS_URL || 'redis://127.0.0.1:6379';
