+++ src/backend/utils/basic_analysis.js
