+++ src/backend/utils/spam-checker.js
