+++ src/backend/utils/text-parser.js
