+++ src/backend/utils/word-counter.js
