+++ test/sanitize-html.test.js
@@ -1,6 +1,6 @@
 // Test for sanitize html
 
-const sanitizeHTML = require('../src/sanitize-html');
+const sanitizeHTML = require('../src/backend/utils/sanitize-html');
 
 test('Tests for sanitized HTML', async () => {
   const data = await sanitizeHTML.run('<img onerror="alert(1)">');
