+++ test/word-counter.test.js
@@ -1,4 +1,4 @@
-const { wordCounter } = require('../src/word-counter');
+const { wordCounter } = require('../src/backend/utils/word-counter');
 
 test('counts the number of words in a 5-word long sentence', () =>
   wordCounter('one two three four five').then(data => {
